I put together a super simple interpreter that I am in the process of rewriting, but I thought I'd commit it for posterity's sake.
The primary purpose of this simple interpreter was to help me organize my thoughts and get through the brain fog. I'm a bit rusty at writing parsers and had to shake off the dust.

I don't plan on using it and will likely delete it, but it'll be in the commit history for us to remember where it started.