Initial pass at layer normalization, improved profiling, L2 Regularization, exit strategy, and improvements to batching -- including a restructuring of the training loop and the backward pass.
I removed mixedPrecisionScale, which may be an issue, but I was able to train 8-bit models still, so I think it is fine. This was always a heavy-handed approach. I think layer normalization and l2 regularization should help with this, but it still may be necessary to tune learning rates for models.